// src/generated/api/model/updateOrganisationDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * NestJS Boilerplate API
 * API documentation for the NestJS boilerplate project.
 * OpenAPI spec version: 1.0
 */

export interface UpdateOrganisationDto {
  name?: string;
  address?: string;
  postalCode?: string;
  city?: string;
  kvk?: string;
  vatNumber?: string;
  iban?: string;
}